requirements.txt
django>=3.0,<4.0
djangorestframework>=3.12.0,<4.0
python-dotenv
django-background-tasks
requests
